Rename main folder scripts to DBS-Pro. Renaming DBSpro python package to dbspro (according to pep8).
