+++ .gitignore
@@ -1,3 +1,3 @@
 __pycache__/
 
-/src/DBSpro.egg-info/
+/src/DBS-Pro.egg-info/
