+++ .travis.yml
@@ -25,9 +25,9 @@ before_install:
   - conda update -q conda
   - conda config --add channels bioconda --add channels conda-forge
   - conda info -a
-  - wget https://export.uppmax.uu.se/uppstore2018173/DBSpro-testdata-0.2.tar.gz
-  - tar xf DBSpro-testdata-0.2.tar.gz
-  - ln -s DBSpro-testdata-0.2 testdata
+  - wget https://export.uppmax.uu.se/uppstore2018173/DBS-Pro-testdata-0.2.tar.gz
+  - tar xf DBS-Pro-testdata-0.2.tar.gz
+  - ln -s DBS-Pro-testdata-0.2 testdata
 
 install:
   - conda env create -n testenv -f environment.yml
